DjangoBackend/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.10.1
